--- src/main/java/com/meli/projetointegradormelifrescos/enums/repository/InboundOrderRepo.java
@@ -1,12 +1,0 @@
-package com.meli.projetointegradormelifrescos.enums.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.meli.projetointegradormelifrescos.model.InboundOrder;
-
-@Repository
-public interface InboundOrderRepo extends JpaRepository<InboundOrder, Long> {
-    InboundOrder save(InboundOrder inboundOrder);
-}
-
