--- src/main/java/com/meli/projetointegradormelifrescos/enums/repository/ManagerRepo.java
@@ -1,11 +1,0 @@
-package com.meli.projetointegradormelifrescos.enums.repository;
-
-import com.meli.projetointegradormelifrescos.model.Manager;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface ManagerRepo extends JpaRepository<Manager, Long> {
-
-    Manager findManagerById(Long managerid);
-
-}
