--- src/main/java/com/meli/projetointegradormelifrescos/enums/repository/SectionRepo.java
@@ -1,7 +1,0 @@
-package com.meli.projetointegradormelifrescos.enums.repository;
-
-import com.meli.projetointegradormelifrescos.model.Section;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SectionRepo extends JpaRepository<Section, Long> {
-}
