+++ src/main/java/com/meli/projetointegradormelifrescos/repository/InboundOrderRepo.java
@@ -11,7 +11,12 @@ package com.meli.projetointegradormelifrescos.repository;
 import com.meli.projetointegradormelifrescos.model.PurchaseProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
- @Repository
-public interface PurchaseProductRepo extends JpaRepository<PurchaseProduct, Long> {
-    PurchaseProduct save(PurchaseProduct purchaseProduct);
+
+import com.meli.projetointegradormelifrescos.model.InboundOrder;
+
+@Repository
+public interface InboundOrderRepo extends JpaRepository<InboundOrder, Long> {
+    InboundOrder save(InboundOrder inboundOrder);
+
 }
+
