+++ src/main/java/com/meli/projetointegradormelifrescos/repository/ManagerRepo.java
@@ -8,10 +8,12 @@
 
 package com.meli.projetointegradormelifrescos.repository;
 
-import com.meli.projetointegradormelifrescos.model.PurchaseProduct;
+import com.meli.projetointegradormelifrescos.model.Manager;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
- @Repository
-public interface PurchaseProductRepo extends JpaRepository<PurchaseProduct, Long> {
-    PurchaseProduct save(PurchaseProduct purchaseProduct);
+
+
+public interface ManagerRepo extends JpaRepository<Manager, Long> {
+
+    Manager findManagerById(Long managerid);
+
 }
