+++ src/main/java/com/meli/projetointegradormelifrescos/repository/SectionRepo.java
@@ -8,10 +8,8 @@
 
 package com.meli.projetointegradormelifrescos.repository;
 
-import com.meli.projetointegradormelifrescos.model.PurchaseProduct;
+import com.meli.projetointegradormelifrescos.model.Section;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
- @Repository
-public interface PurchaseProductRepo extends JpaRepository<PurchaseProduct, Long> {
-    PurchaseProduct save(PurchaseProduct purchaseProduct);
+
+public interface SectionRepo extends JpaRepository<Section, Long> {
 }
