+++ src/main/java/com/meli/projetointegradormelifrescos/repository/WarehouseRepo.java
@@ -8,10 +8,15 @@
 
 package com.meli.projetointegradormelifrescos.repository;
 
-import com.meli.projetointegradormelifrescos.model.PurchaseProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
- @Repository
-public interface PurchaseProductRepo extends JpaRepository<PurchaseProduct, Long> {
-    PurchaseProduct save(PurchaseProduct purchaseProduct);
+import com.meli.projetointegradormelifrescos.model.Warehouse;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface WarehouseRepo extends JpaRepository<Warehouse,Long> {
+    Optional<Warehouse> findWarehouseByCode(Long code);
+    List<Warehouse> findAll();
+
+
 }
