+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.datasource.username=root
-spring.datasource.password=password
+spring.datasource.password=root
 spring.datasource.url=jdbc:mysql://localhost:3306/integrador?createDatabaseIfNotExist=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 spring.mvc.pathmatch.matching-strategy=ant_path_matcher
